Add initial project files for Module 1 homework
Create a new Jupyter notebook for the introduction
to machine learning homework, which includes various
data analysis tasks using Pandas.

Also, add a `.python-version` file to specify the
Python version as 3.11 and a `pyproject.toml` file
to manage dependencies including ipykernel, matplotlib,
numpy, pandas, scikit-learn, and seaborn. This sets
up the environment for the project and ensures that
the necessary packages are available for development.

This environment is created based on the setting up guide provided
in the Zoomcamp course [here](https://github.com/DataTalksClub/machine-learning-zoomcamp/blob/master/01-intro/06-environment.md)