Rename project files and update dependencies
This commit introduces the convention of having a single python-version
and dependencies for a module.

Moved the `.python-version` and `uv.lock` files to the root of the
`01-intro` directory to standardize project structure.

Replaced the existing `pyproject.toml` for the homework with a new
version that reflects the general introduction to the Machine Learning
Zoomcamp. The dependencies remain the same, but the project description
was updated to align with the new context.